+++ core/middleware.py
@@ -28,15 +28,16 @@ class UserSpecificRedirectMiddleware(MiddlewareMixin):
             request.session[self.SESSION_KEY_LEARN] = True
 
 
-class StoreUserInterestsMiddleware(MiddlewareMixin):
+class StoreUserExpertiseMiddleware(MiddlewareMixin):
     def process_request(self, request):
-        if request.user.is_anonymous or 'store' not in request.GET:
+        if request.user.is_anonymous or 'remember-expertise-products-services' not in request.GET:
             return
 
-        products = request.GET.get('products')
+        products = request.GET.getlist('product')
 
-        if products and products != request.user.company.data['expertise_products_services']:
+        if request.user.company and products and products != request.user.company.expertise_products_services:
             helpers.update_company_profile(
                 sso_session_id=request.user.session_id,
-                data={'expertise_products_services': products}
+                data={'expertise_products_services': {'other': products}}
             )
+            request.user.company.data['expertise_products_services']['other'] = products